refactor: add get course detail function for getCourseDetail api logics in courseServices and move queryCourse  to be the  method in courseServices Object
