+++ apps/core/src/routes/courses.ts
@@ -1,5 +1,4 @@
-import { prisma } from "@/db/clients.js";
-import { Prisma, ReviewStatus, VoteType } from "@/generated/prisma/client.js";
+import { Prisma } from "@/generated/prisma/client.js";
 import type { Variables } from "@/lib/auth.js";
 import {
   addFavoriteCourse,
@@ -7,12 +6,9 @@ import {
   getCoursesRoute,
   removeFavoriteCourse,
 } from "@/routes_define/courses.routes.js";
-import { courseServices, queryCourse } from "@/services/coursesService.js";
-import { mapSemester, mapStudyProgram } from "@/utils/enumMapper.js";
+import { courseServices } from "@/services/coursesService.js";
 
 import { OpenAPIHono } from "@hono/zod-openapi";
-
-import type { CourseReview } from "@cugetreg/zod-schemas/courses-response";
 
 import { middlewareAuth } from "./auth.js";
 
@@ -26,7 +22,7 @@ courses
     try {
       const query = c.req.valid("query");
       const user = c.get("user");
-      const result = await queryCourse(query, user?.id);
+      const result = await courseServices.queryCourse(query, user?.id);
       return c.json(result, 200);
     } catch (err) {
       if (err instanceof Error) {
@@ -50,83 +46,23 @@ courses
     try {
       const { courseNo } = c.req.valid("param");
       const { studyProgram, academicYear, semester } = c.req.valid("query");
-
-      const course = await prisma.course.findFirst({
-        where: {
-          courseNo,
-          studyProgram: mapStudyProgram(studyProgram),
-          academicYear,
-          semester: mapSemester(semester),
-        },
-        include: {
-          courseInfo: true,
-          sections: { include: { classes: true } },
-        },
-      });
-
-      if (!course) {
-        return c.json({ message: "Course not found" }, 404);
-      }
-
       const userId = c.get("user")?.id;
 
-      const allReviews = await prisma.review.findMany({
-        where: {
-          courseNo,
-          ...(userId && {
-            OR: [
-              { userId },
-              { status: ReviewStatus.APPROVED, userId: { not: userId } },
-            ],
-          }),
-          ...(!userId && { status: ReviewStatus.APPROVED }),
-        },
-        include: {
-          votes: true,
-        },
-      });
+      const query = { studyProgram, academicYear, semester };
 
-      const reviews = allReviews.map((review) => {
-        let reaction: VoteType | undefined = undefined;
-        const [likeCount, dislikeCount] = review.votes.reduce(
-          ([like, dislike], vote) => {
-            if (vote.userId === userId) {
-              reaction = vote.voteType;
-            }
-
-            return [
-              like + (vote.voteType === VoteType.L ? 1 : 0),
-              dislike + (vote.voteType === VoteType.D ? 1 : 0),
-            ];
-          },
-          [0, 0],
-        );
-
-        return {
-          id: review.id,
-          rating: review.rating,
-          status: review.status,
-          studyProgram: review.studyProgram,
-          academicYear: review.academicYear,
-          semester: review.semester,
-          content: review.content,
-          stats: {
-            likeCount,
-            dislikeCount,
-          },
-          reaction,
-        } as CourseReview;
-      });
-
-      return c.json(
-        {
-          course,
-          reviews,
-        },
-        200,
+      const data = await courseServices.getCourseDetail(
+        query,
+        courseNo,
+        userId,
       );
+      return c.json(data, 200);
     } catch (error) {
       console.error(error);
+      if (error instanceof Error) {
+        if (error.message === "Course not found") {
+          return c.json({ message: "Course not found" }, 404);
+        }
+      }
       return c.json({ error: "INTERNAL_SERVER_ERROR" }, 500);
     }
   })
